feat(Target): add Target class

+++ Assets/StandWorld/Scripts/Helpers/Utils.cs
@@ -6,22 +6,19 @@ namespace StandWorld.Helpers
     {
         public static float Distance(Vector2Int a, Vector2Int b)
         {
-            if (
-                Mathf.Abs(a.x - b.x) + Mathf.Abs(a.y - b.y) == 1
-            ) {
+            if (Mathf.Abs(a.x - b.x) + Mathf.Abs(a.y - b.y) == 1) 
+            {
                 return 1f;
             }
 
-            if (
-                Mathf.Abs(a.x - b.x) == 1 && 
-                Mathf.Abs(a.y - b.y) == 1
-            ) {
+            if (Mathf.Abs(a.x - b.x) == 1 && Mathf.Abs(a.y - b.y) == 1)
+            {
                 return 1.41121356237f;
             }
 
             return Mathf.Sqrt(
-                Mathf.Pow((float)a.x-(float)b.x, 2) +
-                Mathf.Pow((float)a.y-(float)b.y, 2)				
+                Mathf.Pow(a.x - (float)b.x, 2) +
+                Mathf.Pow(a.y - (float)b.y, 2)				
             );
         }
     }
